[refactor] Refatorando código para armazenamento de dados da Conta

+++ src/main/kotlin/com/rcrdev/instituicao/InstituicaoRepository.kt
@@ -4,5 +4,5 @@ import io.micronaut.data.annotation.Repository
 import io.micronaut.data.jpa.repository.JpaRepository
 
 @Repository
-interface InstituicaoRepository: JpaRepository<Instituicao, Long> {
+interface InstituicaoRepository: JpaRepository<Instituicao, Int> {
 }
